refactor: extract shared language infrastructure into src/lang/
Move tree-sitter grammars, file type detection, outline extraction,
and generated/binary detection from read/ and search/ into a new
lang/ foundation module. All feature modules (read, search, index,
map, edit, diff) now import shared infrastructure from one place.

- lang/mod.rs: detect_file_type(), package_root()
- lang/outline.rs: outline_language(), walk_top_level(), get_outline_entries()
- lang/treesitter.rs: DEFINITION_KINDS, extract_definition_name(), definition_weight()
- lang/detection.rs: is_generated_by_name/content(), is_binary()
- format.rs: rel() moved from search/mod.rs (now pub(crate))
- Deleted: read/binary.rs, read/generated.rs, search/treesitter.rs

Zero behavior changes. 151 tests pass, 0 clippy warnings.
