+++ src/read/mod.rs
@@ -1,5 +1,3 @@
-pub mod binary;
-pub mod generated;
 pub mod imports;
 pub mod outline;
 
@@ -11,7 +9,8 @@ use memmap2::Mmap;
 use crate::cache::OutlineCache;
 use crate::error::TilthError;
 use crate::format;
-use crate::types::{estimate_tokens, FileType, Lang, ViewMode};
+use crate::lang::detect_file_type;
+use crate::types::{estimate_tokens, FileType, ViewMode};
 
 pub(crate) const TOKEN_THRESHOLD: u64 = 6_000;
 const FILE_SIZE_CAP: u64 = 500_000; // 500KB
@@ -84,7 +83,7 @@ pub fn read_file(
     })?;
     let buf = &mmap[..];
 
-    if binary::is_binary(buf) {
+    if crate::lang::detection::is_binary(buf) {
         let mime = mime_from_ext(path);
         return Ok(format::binary_header(path, byte_len, mime));
     }
@@ -92,7 +91,9 @@ pub fn read_file(
     let name = path.file_name().and_then(|n| n.to_str()).unwrap_or("");
 
     // Generated
-    if generated::is_generated_by_name(name) || generated::is_generated_by_content(buf) {
+    if crate::lang::detection::is_generated_by_name(name)
+        || crate::lang::detection::is_generated_by_content(buf)
+    {
         let line_count = memchr::memchr_iter(b'\n', buf).count() as u32 + 1;
         return Ok(format::file_header(
             path,
@@ -367,45 +368,6 @@ fn list_directory(path: &Path) -> Result<String, TilthError> {
 
     let header = format!("# {} ({} items)", path.display(), items.len());
     Ok(format!("{header}\n\n{}", entries.join("\n")))
-}
-
-/// Detect file type by extension, then by name.
-pub fn detect_file_type(path: &Path) -> FileType {
-    match path.extension().and_then(|e| e.to_str()) {
-        Some("ts") => FileType::Code(Lang::TypeScript),
-        Some("tsx") => FileType::Code(Lang::Tsx),
-        Some("js" | "jsx") => FileType::Code(Lang::JavaScript),
-        Some("py" | "pyi") => FileType::Code(Lang::Python),
-        Some("rs") => FileType::Code(Lang::Rust),
-        Some("go") => FileType::Code(Lang::Go),
-        Some("java") => FileType::Code(Lang::Java),
-        Some("scala" | "sc") => FileType::Code(Lang::Scala),
-        Some("c" | "h") => FileType::Code(Lang::C),
-        Some("cpp" | "hpp" | "cc" | "cxx") => FileType::Code(Lang::Cpp),
-        Some("rb") => FileType::Code(Lang::Ruby),
-        Some("php" | "phtml") => FileType::Code(Lang::Php),
-        Some("swift") => FileType::Code(Lang::Swift),
-        Some("kt" | "kts") => FileType::Code(Lang::Kotlin),
-        Some("cs") => FileType::Code(Lang::CSharp),
-
-        Some("md" | "mdx" | "rst") => FileType::Markdown,
-        Some("json" | "yaml" | "yml" | "toml" | "xml" | "ini") => FileType::StructuredData,
-        Some("csv" | "tsv") => FileType::Tabular,
-        Some("log") => FileType::Log,
-
-        None => file_type_from_name(path),
-        _ => FileType::Other,
-    }
-}
-
-fn file_type_from_name(path: &Path) -> FileType {
-    match path.file_name().and_then(|n| n.to_str()) {
-        Some("Dockerfile" | "Containerfile") => FileType::Code(Lang::Dockerfile),
-        Some("Makefile" | "GNUmakefile") => FileType::Code(Lang::Make),
-        Some("Vagrantfile" | "Rakefile") => FileType::Code(Lang::Ruby),
-        Some(n) if n.starts_with(".env") => FileType::StructuredData,
-        _ => FileType::Other,
-    }
 }
 
 /// Public entry point for did-you-mean on path-like fallthrough queries.
